% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_values}
\alias{example_values}
\title{Randomize values for a variable.}
\usage{
example_values(x, max_value = 100)
}
\arguments{
\item{x}{Linear problem variable. Get it from 'easylp$variables'.}

\item{max_value}{Numeric, maximum absolute value of the values, when
the variable is unbounded.}
}
\value{
An array of the same dimensions as the variable.
}
\description{
You can use this function to test if the variable is correctly defined,
or to check how you may index it. The values aren't necessarily feasable,
but they are bounded. If the variable is integer or binary, the values will too.
}
