% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easylp.R
\name{lp_con}
\alias{lp_con}
\title{Define a constraint for a linear problem.}
\usage{
lp_con(constraint, ...)
}
\arguments{
\item{constraint}{In the form of \code{2*x - y <= b}. The variables
must always be on the left hand side. The inequality can be one of:
'<', '<=', '==', '>', '>='.}

\item{...}{Abbreviations for variables. See \link{examples}.}
}
\value{
Object of class 'lp_con'
}
\description{
Creates one or more linear constraints
}
\details{
If the left hand side of the constraint is not atomic, the values are summed.
For example, \code{lp_con(x[1,] <= 10)} means 'the sum of the first row
of x is less than 10.'

If you want the constraint to be 'all values in the first row of x must
be less than 10' you need to use this syntax:
\code{lp_con( for (j in 1:ncol(x)) x[1, j] <= 10)}.
}
\examples{
product <- c("apple", "orange", "lemon")
stock <- c(apple = 150, orange = 120, lemon = 60)
x <- lp_var("sales", product = product, integer = TRUE)

# Max stock for each product
max_stock <- lp_con( for (p in product[1:2]) x[p] <= stock[p])
max_stock

# 'apple' + 'orange' >= 240
min_total <- lp_con(x[p] >= 240, p = product[1:2])
min_total

# Notice how 'p' serves very different purposes in the two cases.
}
