% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easylp.R
\name{lp_var}
\alias{lp_var}
\title{Define a variable for a linear problem.}
\usage{
lp_var(name, ..., bound = NULL, integer = FALSE, binary = FALSE)
}
\arguments{
\item{...}{Sets where the variables lives. Must be \link{lp_set} objects.
The order is important, as it defines the shape of the array:
\code{dim(x) = c(length(set_1), length(set_2), ...)}}

\item{bound}{Inequality of the shape \code{.x >= 0, .x <= 2},
or \code{NULL} if it's free.}

\item{integer}{Is the variable an integer?}

\item{binary}{Is the variable binary {0, 1}?}
}
\value{
Object of class 'lp_var'
}
\description{
Creates an object of class 'lp_var' to be used in lp_solve.
}
\examples{
# To do
}
