% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter}
\alias{parameter}
\title{Define a parameter for a linear problem.}
\usage{
parameter(x, ..., byrow = FALSE)
}
\arguments{
\item{x}{Coefficients for the parameter.}

\item{...}{Sets to index the parameter.}
}
\value{
Named array.
}
\description{
Automatically set the dimensions and names of a parameter, based on
sets. Similar to \code{number} in the SAS/OR proc optmodel.
}
