% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_for}
\alias{sum_for}
\title{Sum with indexing variables}
\usage{
sum_for(..., .env = caller_env())
}
\arguments{
\item{...}{Named arguments are interpreted as indexing variables. One
unnamed argument should the expression.}

\item{.env}{Environment where the expressions should be evaluated.}
}
\value{
The sum of all evaluated expressions.
}
\description{
Calculates results independently and adds them.
}
\examples{
mat <- matrix(rpois(12, 2), nrow = 3, ncol = 4)
vec <- rpois(4, 5)
sum(mat \%*\% vec)
sum_for(i=1:3, mat[i, ] * vec)
sum_for(i=1:3, j=1:4, mat[i, j] * vec[j])
}
