% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{easylp}
\alias{easylp}
\title{Easy Linear Problem}
\description{
Object containing all the information about a linear problem, as well
as functions to define, modify, and solve it.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{variables}}{List of variables included.}

\item{\code{nvar}}{Actual number of variables, \code{sum(lengths(variables))}.}

\item{\code{constraint}}{List including the constraint matrix 'mat',
a vector of directions 'dir', and a vector of right-hand-side values 'rhs'.}

\item{\code{objective_fun}}{Vector of coefficients for the objective function.}

\item{\code{objective_add}}{Optional value to add to the objective value.}

\item{\code{direction}}{Character indicating whether to minimize 'min' or maximize
'max' the objective function.}

\item{\code{status}}{Character indicating the status of the problem. Initialized as
'unsolved'. Changes when solving the problem using easylp$solve().}

\item{\code{objective_value}}{Value of the objective function with the optimal solution.}

\item{\code{solution}}{Optimal values of the variables.}

\item{\code{pointer}}{Pointer to an lpSolveAPI object.
See more at \url{https://lpsolve.sourceforge.net/5.5/}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-easylp-var_}{\code{easylp$var_()}}
\item \href{#method-easylp-var}{\code{easylp$var()}}
\item \href{#method-easylp-con}{\code{easylp$con()}}
\item \href{#method-easylp-min}{\code{easylp$min()}}
\item \href{#method-easylp-max}{\code{easylp$max()}}
\item \href{#method-easylp-solve_}{\code{easylp$solve_()}}
\item \href{#method-easylp-solve}{\code{easylp$solve()}}
\item \href{#method-easylp-feasable}{\code{easylp$feasable()}}
\item \href{#method-easylp-check_feasable}{\code{easylp$check_feasable()}}
\item \href{#method-easylp-optimal}{\code{easylp$optimal()}}
\item \href{#method-easylp-check_optimal}{\code{easylp$check_optimal()}}
\item \href{#method-easylp-.obj}{\code{easylp$.obj()}}
\item \href{#method-easylp-.test}{\code{easylp$.test()}}
\item \href{#method-easylp-pretty_solution}{\code{easylp$pretty_solution()}}
\item \href{#method-easylp-print}{\code{easylp$print()}}
\item \href{#method-easylp-clone}{\code{easylp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-var_"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-var_}{}}}
\subsection{Method \code{var_()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$var_(name, ..., integer = FALSE, binary = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-var"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-var}{}}}
\subsection{Method \code{var()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$var(definition, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-con"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-con}{}}}
\subsection{Method \code{con()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$con(..., expr_list = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-min"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-min}{}}}
\subsection{Method \code{min()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$min(objective)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-max"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-max}{}}}
\subsection{Method \code{max()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$max(objective)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-solve_"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-solve_}{}}}
\subsection{Method \code{solve_()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$solve_(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-solve"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-solve}{}}}
\subsection{Method \code{solve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$solve()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-feasable"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-feasable}{}}}
\subsection{Method \code{feasable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$feasable(solution = self$solution, tol = 2e-08)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-check_feasable"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-check_feasable}{}}}
\subsection{Method \code{check_feasable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$check_feasable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-optimal"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-optimal}{}}}
\subsection{Method \code{optimal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$optimal(solution, base, tol = 2e-08)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-check_optimal"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-check_optimal}{}}}
\subsection{Method \code{check_optimal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$check_optimal(solution = self$solution, tol = 2e-08, max_iter = 50L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-.obj"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-.obj}{}}}
\subsection{Method \code{.obj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$.obj(expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-.test"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-.test}{}}}
\subsection{Method \code{.test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$.test(expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-pretty_solution"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-pretty_solution}{}}}
\subsection{Method \code{pretty_solution()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$pretty_solution()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-print"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-easylp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-easylp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{easylp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
