% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter}
\alias{parameter}
\title{Define a parameter for a linear problem.}
\usage{
parameter(x, ..., byrow = FALSE)
}
\arguments{
\item{x}{Coefficients for the parameter.}

\item{...}{Sets to index the parameter.}

\item{byrow}{If there are 2 sets in \code{...}, whether to fill
the matrix by rows. Otherwise filled by columns}
}
\value{
Named array.
}
\description{
Automatically set the dimensions and names of a parameter, based on
sets. Similar to \code{number} in the SAS/OR proc optmodel.
}
\examples{
factory <- c("A", "B")
market <- c(1:3)
transport_cost <- c(
    3, 4, 2,
    6, 2, 5
) |> parameter(factory, market, byrow = TRUE)
}
